+++ lisp/core/compile.el
@@ -56,6 +56,7 @@
   (let* ((compiled (cl-remove-if-not #'eask--byte-compile-file files))
          (compiled (length compiled))
          (skipped (- (length files) compiled)))
+    (eask-msg "")
     (eask-info "(Total of %s file%s compiled, %s skipped)" compiled
                (eask--sinr compiled "" "s")
                skipped)))
@@ -66,6 +67,6 @@
                       (eask-package-el-files))))
       (eask--compile-files files)
     (eask-info "(No files have been compiled)")
-    (eask-help 'compile)))
+    (eask-help "core/compile")))
 
 ;;; core/compile.el ends here