+++ lisp/core/load-path.el
@@ -21,6 +21,7 @@
 (eask-start
   (eask-pkg-init)
   (mapc #'eask--print-load-path load-path)
+  (eask-msg "")
   (eask-info "(Total of %s load-path)" (length load-path)))
 
 ;;; core/load-path.el ends here