+++ lisp/core/recipe.el
@@ -34,7 +34,8 @@
         (eask-msg "")
         (eask-msg "Recipe: %s" (pp-to-string recipe))
         (eask-msg ""))
+    (eask-msg "")
     (eask-info "(Repository URL is required to form a recipe)")
-    (eask-help 'recipe)))
+    (eask-help "core/recipe")))
 
 ;;; core/recipe.el ends here