+++ lisp/core/reinstall.el
@@ -41,8 +41,10 @@
               ((package-installed-p name)))
         (progn
           (eask-package-reinstall name)
+          (eask-msg "")
           (eask-info "(Reinstalled %s)" name))
+      (eask-msg "")
       (eask-info "✗ (No packages have been reintalled)")
-      (eask-help 'reinstall))))
+      (eask-help "core/reinstall"))))
 
 ;;; core/reinstall.el ends here