+++ lisp/core/search.el
@@ -37,9 +37,11 @@
           (setq result (append result (eask--search-packages query))))
         (delete-dups result)
         (eask--list result package-archive-contents)
+        (eask-msg "")
         (eask-info "(Search result of %s package%s)" (length result)
                    (eask--sinr result "" "s")))
+    (eask-msg "")
     (eask-info "(No search operation; missing queries specification)")
-    (eask-help 'search)))
+    (eask-help "core/search")))
 
 ;;; core/search.el ends here